Add ability to add a node into a cluster
A new node can be added into a cluster. The node will be provisioned
from a cloud provider (Jetstream IU for the time being) and then
configured (via an Ansible playbook). Conceptually, the node is added
by going through 'provision' and 'configure' steps. The steps are
intended to be configurable by supplying desired managers.

This code requires CloudBridge at revision
https://github.com/gvlproject/cloudbridge/commit/9b19e2d6c803f8b17dec16032bdbc4fe01ace248
or later.